+++ ZXingObjC/client/result/ZXURLTOResultParser.m
@@ -25,7 +25,7 @@
   if (![rawText hasPrefix:@"urlto:"] && ![rawText hasPrefix:@"URLTO:"]) {
     return nil;
   }
-  int titleEnd = [rawText rangeOfString:@":" options:NSLiteralSearch range:NSMakeRange(6, [rawText length] - 6)].location;
+  NSUInteger titleEnd = [rawText rangeOfString:@":" options:NSLiteralSearch range:NSMakeRange(6, [rawText length] - 6)].location;
   if (titleEnd == NSNotFound) {
     return nil;
   }
