+++ ZXingObjCTests/oned/ZXEAN13WriterTestCase.m
@@ -24,7 +24,7 @@
   NSString *testStr = @"00010100010110100111011001100100110111101001110101010110011011011001000010101110010011101000100101000";
   ZXBitMatrix *result = [[[ZXEAN13Writer alloc] init] encode:@"5901234123457"
                                                       format:kBarcodeFormatEan13
-                                                       width:testStr.length
+                                                       width:(int)testStr.length
                                                       height:0
                                                        error:nil];
   for (int i = 0; i < testStr.length; i++) {
