+++ ZXingObjCTests/oned/ZXEAN8WriterTestCase.m
@@ -24,7 +24,7 @@
   NSString *testStr = @"0001010001011010111101111010110111010101001110111001010001001011100101000";
   ZXBitMatrix *result = [[[ZXEAN8Writer alloc] init] encode:@"96385074"
                                                      format:kBarcodeFormatEan8
-                                                      width:testStr.length
+                                                      width:(int)testStr.length
                                                      height:0
                                                       error:nil];
   for (int i = 0; i < testStr.length; i++) {
